Remove Boost.Interprocess code and avoid server crashes
Shared memory is overkill if all we need is to wait for a few seconds until server starts.
It's also not that good for Android and caused problems on Windows too.

Now server not going to crash if you connect to it using netcat and then disconnect.
Also server now report it's state when it's waiting for connection or in pregame.