Fix lag on dragging mod list in Launcher on Android
Tested by enabling dragging on desktop. According to profiler, most of
time was spent on generating mod information. Cheap to do once, but it
looks like Qt does this A LOT.

Added simple caching & reduced copying. Seems to be working fine on PC,
but not yet confirmed whether same goes for mobile